backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
anthropic==0.42.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
